+++ web-vue/vite.config.ts
@@ -8,6 +8,7 @@ export default defineConfig(({ command }) => ({
     outDir: '../static/vue-app',
     emptyOutDir: true,
     sourcemap: false,
+    chunkSizeWarningLimit: 900,
     rollupOptions: {
       output: {
         manualChunks: {